Delete existing attachments before WordPress checks for naming conflict.
This commit changes the approach that the plugin takes to solving the problem of conflicting uploads. Previously the plugin required a Core hack so it could filter the unique filename that WordPress generates. That sucked, for obvious reasons.

This approach is more elegant, and works by deleting the old file before WordPress checks for it.


git-svn-id: https://plugins.svn.wordpress.org/overwrite-uploads/trunk@822228 b8457f37-d9ea-0310-8a92-e5e31aec5664